requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pika==1.1.0
psycopg2==2.8.5
pyaml==20.4.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
sqlparse==0.3.1
threadpoolctl==2.0.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
